Create feature spec helper
